requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
PyMySQL==1.1.1
Pillow==10.4.0
requests==2.32.3